+++ lib/src/core/presenter.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/widgets.dart';
+
+mixin Presenter {
+  Widget buildPresenter(BuildContext context);
+}