+++ lib/src/core/uri.dart
@@ -73,9 +73,11 @@ typedef UriMapBuilder<Out, State> = Out? Function(UriMatch<State> match);
 
 /// Matched values & context provided to a [UriMapBuilder].
 class UriMatch<State> {
-  UriMatch(this.uri,
-      this.pathParameters,
-      this.state,);
+  UriMatch(
+    this.uri,
+    this.pathParameters,
+    this.state,
+  );
 
   final Uri uri;
   final Map<String, String> pathParameters;
@@ -86,16 +88,18 @@ class UriMatch<State> {
 
 /// Associates one (or many) path patterns with a builder.
 class UriMap<Out, State> {
-  UriMap(String pattern,
-      this.builder, {
-        bool matchEnd = true,
-      }) : matchers = [PathMatcher(pattern, matchEnd: matchEnd)];
+  UriMap(
+    String pattern,
+    this.builder, {
+    bool matchEnd = true,
+  }) : matchers = [PathMatcher(pattern, matchEnd: matchEnd)];
 
-  UriMap.many(List<String> patterns,
-      this.builder, {
-        bool matchEnd = true,
-      }) : matchers =
-  patterns.map((e) => PathMatcher(e, matchEnd: matchEnd)).toList();
+  UriMap.many(
+    List<String> patterns,
+    this.builder, {
+    bool matchEnd = true,
+  }) : matchers =
+            patterns.map((e) => PathMatcher(e, matchEnd: matchEnd)).toList();
 
   final List<PathMatcher> matchers;
   final UriMapBuilder builder;
@@ -111,7 +115,8 @@ class UriMap<Out, State> {
 /// * - anything
 /// Compiles a path pattern with named fields into a matching regex.
 class PathMatcher {
-  PathMatcher(this.pattern, {
+  PathMatcher(
+    this.pattern, {
     // match end [$]
     bool matchEnd = true,
   }) {
@@ -179,7 +184,7 @@ class PathMatcher {
     }
     final map = fields.fold(
       <String, String>{},
-          (p, e) => p..[e] = match.namedGroup(e)!,
+      (p, e) => p..[e] = match.namedGroup(e)!,
     );
     return fields.every(map.containsKey) ? map : null;
   }